misunderstandings

+++ anreader_sources/device/n3alarms_settings.h
@@ -56,8 +56,8 @@ public:
     unsigned interval(int index) const;
     unsigned max(int index) const;
     unsigned min(int index) const;
-    unsigned meters2feet(int value) const;
-    unsigned feet2meters(int value, alarm_type atype) const;
+    double meters2feet(double value) const;
+    double feet2meters(double value, alarm_type atype) const;
 
 };
